requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain==0.1.9
langchain-google-genai==0.0.11
google-generativeai>=0.4.1
python-multipart==0.0.9
pydantic==2.6.1
python-jose==3.3.0
redis==5.0.1
sqlalchemy==2.0.27
aiofiles==23.2.1
python-dotenv==1.0.1
pypdf==4.0.1
email-validator==2.1.0.post1
httpx==0.26.0
jsonschema==4.21.1 